feat: Persona가 삭제되면 길드에 저장된 대표펫이 자동으로 변경되도록 수정한다

+++ src/main/kotlin/org/gitanimals/guild/app/GetJoinedGuildFacade.kt
@@ -13,6 +13,6 @@ class GetJoinedGuildFacade(
     fun getJoinedGuilds(token: String): List<Guild> {
         val user = identityApi.getUserByToken(token)
 
-        return guildService.findAllGuildByUserId(user.id)
+        return guildService.findAllGuildByUserId(user.id.toLong())
     }
 }
